oasparser/requirements.txt
Flask==2.2.2
PyYAML==5.4.1
Werkzeug==2.2.2 